fix warnings, commenting off 'find_path' and 'find_library' for Mercury in src/commons/CMakeLists.txt  (#93)
* Revert "update output"

This reverts commit fe1f8b44995bc0dabd3b957e1032c2da26f56fdd.

* build kvtag_add_get_scale

* comment off free

* update code

* 1. kvtag_scale_add_get added \n 2. uint64_t support for obj/tag/query count \n 3. moving work assigning block downwards right before creating objects \n 4. everything is tested working

* do while loop added, tested with 1m object and works

* 1m objects test works, 10m object test fail as the original also fails

* add new executable to test set

* enlarge PDC_SERVER_ID_INTERVAL

* update code

* update console args

* add p search test

* add console arg for changing number of attributes per object

* free allocated memory

* fix query count issue

* fix attr length definition

* code refactored

* code refactored

* code refactored

* code refactored

* code refactored

* code refactored

* fix data type

* fix data type

* fix data type

* add client side statistics

* add client side statistics

* fix format

* clang formatter

* update CMake

* update CMake

* update CMake

* free allocated memory properly

* clang format

* clang format

* clang-format-10

* change file name

* address review comments

* update llsm importer

* update llsm importer

* update server checkpoint intervals

* update gitignore

* adding job scripts

* adding one debugging msg

* update container creation to collective mode for debugging purpose

* update container creation to collective mode for debugging purpose

* update container creation to collective mode for debugging purpose

* update container creation to collective mode for debugging purpose

* update output for uint64_t

* add scripts

* update output for uint64_t

* update output for uint64_t

* update output for uint64_t

* update scripts

* update scripts

* delete debugging message

* make Cmake to publish scripts directory

* make Cmake to publish scripts directory

* make Cmake to publish scripts directory

* make Cmake to publish scripts directory

* make Cmake to publish scripts directory

* update tag names

* update tag names

* update query startingpos

* update query startingpos

* update job scripts

* add progressive timing for kvtag_add_get_scale

* fix iteration count in final report

* update job scripts and benckmark program

* update message format

* update message format

* update message format

* update message format

* clang format

* update job scripts

* comment off object/container close procedure in  benchmark to save node hours

* change the max number of object to 1M

* change the max length of attribute value

* change the max length of attribute value

* llsm tiff import test

* llsm tiff import test

* llsm tiff import test

* llsm tiff import test

* update code

* update code

* update code

* update code

* update code

* update code

* update code

* update code

* update code

* update code

* update code

* update cmake and llsm_importer

* update cmake and llsm_importer

* close if in cmake

* cmake fix tiff

* cmake policy to suppress warning

* add pdc include dir

* update code

* update code

* update code

* update code

* update code

* update code

* update array generating method

* update array generating method

* update array generating method

* update array generating method

* update CMakeLists

* update CMakeLists

* update CMakeLists

* update CMakeLists

* update CMakeLists

* fix return type

* fix return type

* add timing

* add timing

* fix output

* llsm tiff importer 1st version: read csv and import tiff files to PDC, adding metadata available in CSV files and TIFF loader

* fix vairable name

* fix cmake

* fix cmake

* fix cmake

* fix cmake

* fix cmake

* fix cmake

* add scripts

* add scripts

* add scripts

* debugging for nonMPI program

* debugging for nonMPI program

* debugging for nonMPI program

* clang format, without PDC, everything works perfectly. program fails at PDC init stage where PDCprop_create(PDC_CONT_CREATE, pdc) is being created

* enable MPI

* enable MPI

* enlarge BCase size

* enlarge BCase size

* enlarge BCase size

* resolve bcast count

* llsm data path in script

* llsm data path in script

* update csv reader

* update csv reader

* update csv reader

* update pdc

* update pdc

* update pdc

* update pdc

* update pdc

* update pdc

* update pdc

* update pdc

* update pdc

* update pdc

* update pdc

* enlarge max write

* update pdc

* update pdc

* update pdc

* update pdc

* update pdc_import.c

* update pdc_import.c

* update pdc_export.c

* update pdc_import.c

* update pdc_import.c

* update pdc_import.c

* update pdc_import.c

* update tools/cmake

* clang format

* clang format

* added a tutorial for llsm_importer

* added a tutorial for llsm_importer

* make sure the line feed is included for string attribute

* update timing for overall completion time

* update formatting

* update metrics

* forcibly enable openmp

* adding C flags from the mex compiler

* Update .gitlab-ci.yml

* updated code

* clang format

* llsm_importer (#1)

formatter on llsm_importer

* add type for kvtag structure (#2)

* upate metadata type system

* update serde framework to coupe with the new data type system

* replace unnecessary data types

* adding type for pdc_kvtag_t, all occurances are fixed

* update new commons CMake for publishing commons

* commons compilation passed

* compiled

* Feature/metadata type (#3)

* upate metadata type system

* update serde framework to coupe with the new data type system

* replace unnecessary data types

* adding type for pdc_kvtag_t, all occurances are fixed

* update new commons CMake for publishing commons

* commons compilation passed

* compiled

* remove unnecessary header files from installation

* resolve conflict

* add important files

* clang formatting

* update cmake

* update

* print numWorkers

* update scripts

* update script

* update script

* formatting

* update llsm_tools.c

* remove unnecessary hash table init

* update script

* fix some warnings

* fix some warnings

* update

* update

* fix warning

* update

* update

* update

* update

* update

* update

* update

* update

* update

* fix warnings

* fix warnings

* fix warnings

* fix warnings

* fix warnings

* fix warnings

* fix warnings

* update

* update

* update

* server address and file paths using 1024, TMP_DIR path using 1024/2, NAME_MAX for appname and objname takes 1024/2, HOSTNAME takes 1024/8, NA_INFO_STRING takes 1024/4

* update

* update

* update

* update

---------

+++ src/tests/vpicio_mts.c
@@ -76,7 +76,7 @@ main(int argc, char **argv)
     uint64_t *offset_remote;
     uint64_t *mysize;
     double    t0, t1;
-    int       steps = 1, sleeptime = 0;
+    uint64_t  steps = 1, sleeptime = 0;
 
     pdcid_t transfer_request_x, transfer_request_y, transfer_request_z, transfer_request_px,
         transfer_request_py, transfer_request_pz, transfer_request_id1, transfer_request_id2;
@@ -95,8 +95,8 @@ main(int argc, char **argv)
         sleeptime    = atoi(argv[3]);
     }
     if (rank == 0)
-        printf("Writing %" PRIu64 " number of particles for %d steps with %d clients.\n", numparticles, steps,
-               size);
+        printf("Writing %" PRIu64 " number of particles for %ld steps with %d clients.\n", numparticles,
+               steps, size);
 
     dims[0] = numparticles;
 
@@ -199,7 +199,7 @@ main(int argc, char **argv)
         MPI_Barrier(MPI_COMM_WORLD);
         t0 = MPI_Wtime();
         if (rank == 0) {
-            printf("\n#Step  %llu\n", i);
+            printf("\n#Step  %lu\n", i);
         }
 #endif
         PDCprop_set_obj_time_step(obj_prop_xx, i);
@@ -504,7 +504,7 @@ main(int argc, char **argv)
         if (i != steps - 1) {
             sleep(sleeptime);
             if (rank == 0) {
-                printf("Sleep time: %d.00\n", sleeptime);
+                printf("Sleep time: %ld.00\n", sleeptime);
             }
         }
     } // End for steps
